djangoBackend/requirements.txt
coreapi==2.3.3
Django==2.1.7
django-allauth==0.39.1
django-dotenv==1.4.2
django-extensions==2.1.4
django-rest-auth==0.9.5
djangorestframework==3.9.0
